refactor: recive progress signal from package manager
When dbus client send a long time request to server, the server must
return an object path as job path, and the client with monitor the
ProgressChanged and Finish signal.

use Job::setProgress and Job::setFinish to control job progress and
finish the job. DO NOT emit job dbus signal direct if you do not
know what will happen. The qt dbus signal MUST be emited on main thread.

Change-Id: I729b8b715d8c3e25f05cbb6368cfbbceb1a747f8